+++ sports/__init__.py
@@ -1,2 +1,3 @@
+from .constants import *
 from .scores import get_all_matches, get_match_score, get_sport_scores
 from .teams import get_team_info
