+++ sports/constants.py
@@ -1,0 +1,9 @@
+BASEBALL = 'baseball'
+BASKETBALL = 'basketball'
+CRICKET = 'cricket'
+FOOTBALL = 'football'
+HANDBALL = 'handball'
+HOCKEY = 'hockey'
+SOCCER = 'soccer'
+TENNIS = 'tennis'
+VOLLEYBALL = 'volleyball'
